refactor(mainapp): extract common upload app logic
